Avoid split packages for Agroal deployment/spi
Might break some of the external extensions so doing it at the beginning
of a new cycle.

Fixes #11064
